fix: calculate average price only when increase balance

+++ src/services/AccountService.ts
@@ -5,8 +5,8 @@ import { Container, Service, Inject } from 'typedi'
 import { lcd } from 'lib/terra'
 import { num } from 'lib/num'
 import { getTokenBalance } from 'lib/mirror'
-import { AssetBalance } from 'graphql/schema'
 import { AssetService } from 'services'
+import { AssetBalance } from 'graphql/schema'
 import { BalanceEntity } from 'orm'
 
 @Service()
@@ -58,16 +58,34 @@ export class AccountService {
     }
 
     const totalBalance = num(entity.balance).plus(amount)
+
+    if (num(price).isGreaterThan(0)) {
+      // average = (last.avg_price*last.amount + current.avg_price*current.amount) / total_amount
+      const value = num(price).multipliedBy(amount)
+      const lastValue = num(entity.averagePrice).multipliedBy(entity.balance)
+
+      entity.averagePrice = lastValue.plus(value).dividedBy(totalBalance).toString()
+    }
+
+    entity.balance = totalBalance.toString()
+
+    return repo.save(entity)
+  }
+
+  async removeBalance(
+    address: string, token: string, amount: string, repo = this.balanceRepo
+  ): Promise<BalanceEntity> {
+    const entity = await this.getBalanceEntity({ address, token }, undefined, repo)
+    if (!entity) {
+      return
+    }
+
+    const totalBalance = num(entity.balance).minus(amount)
     if (totalBalance.isLessThanOrEqualTo(0)) {
       await repo.remove(entity)
       return
     }
 
-    // average = (last.avg_price*last.amount + current.avg_price*current.amount) / total_amount
-    const value = num(price).multipliedBy(amount)
-    const lastValue = num(entity.averagePrice).multipliedBy(entity.balance)
-
-    entity.averagePrice = value.plus(lastValue).dividedBy(totalBalance).toString()
     entity.balance = totalBalance.toString()
 
     return repo.save(entity)
